+++ CHANGELOG.md
@@ -508,3 +508,4 @@
 * Greatly simplified RpgDefaultValues, mostly fixed with no skill or stat definitions
 * Removed most stat and skill definitions from builtin StatType and SkillType enums
 * GameParams will now be loaded from textasset at Data/RPGDefs/rpg_params in addition to external override file
+* Scroll speed in the UI is now multiplied by UIScrollSpeed in config